Created Significand Left Shifter
Initially Shifts whatever the data input is by 2 bits to the right. All
internal signals are 2 extra bits wide to include this buffer space.
This allows each "on" bit to perform a left shift, rather than worrying
about which way the bits should move. Additionally, the final output is
(of 55 downto 0), 54 downto 2, since the MSB is an implicit value in the
mantissa and the 2 LSB were simply buffer space in case the final shift
is to the left overall.